requirements.txt
dspy
nltk
pydantic
python-dotenv
openai
numpy